+++ packages/cli/src/telemetry/index.ts
@@ -10,7 +10,7 @@ import { LicenseService } from '@/license/License.service';
 import { N8N_VERSION } from '@/constants';
 import { SourceControlPreferencesService } from '../environments/sourceControl/sourceControlPreferences.service.ee';
 import { InstanceSettings } from 'n8n-core';
-import { RoleService } from '@/services/role.service';
+import { RoleRepository } from '@/databases/repositories/role.repository';
 
 type ExecutionTrackDataKey = 'manual_error' | 'manual_success' | 'prod_error' | 'prod_success';
 
@@ -109,7 +109,7 @@ export class Telemetry {
 			plan_name_current: this.license.getPlanName(),
 			quota: this.license.getTriggerLimit(),
 			usage: await LicenseService.getActiveTriggerCount(),
-			role_count: await Container.get(RoleService).countUsersByRole(),
+			role_count: await Container.get(RoleRepository).countUsersByRole(),
 			source_control_set_up: Container.get(SourceControlPreferencesService).isSourceControlSetup(),
 			branchName: sourceControlPreferences.branchName,
 			read_only_instance: sourceControlPreferences.branchReadOnly,
