+++ packages/cli/test/integration/role.repository.test.ts
@@ -1,28 +1,17 @@
-import { RoleService } from '@/services/role.service';
 import { createAdmin, createMember, createOwner } from './shared/db/users';
 import * as testDb from './shared/testDb';
-import { mock } from 'jest-mock-extended';
 import { RoleRepository } from '@/databases/repositories/role.repository';
 import Container from 'typedi';
-import { SharedWorkflowRepository } from '@/databases/repositories/sharedWorkflow.repository';
 
-describe('RoleService', () => {
-	let roleService: RoleService;
+describe('RoleRepository', () => {
+	let roleRepository: RoleRepository;
 
 	beforeAll(async () => {
 		await testDb.init();
 
-		roleService = new RoleService(
-			Container.get(RoleRepository),
-			Container.get(SharedWorkflowRepository),
-			mock(),
-		);
+		roleRepository = Container.get(RoleRepository);
 
 		await testDb.truncate(['User']);
-	});
-
-	afterEach(async () => {
-		await testDb.truncate(['SharedWorkflow']);
 	});
 
 	afterAll(async () => {
@@ -40,7 +29,7 @@ describe('RoleService', () => {
 				createMember(),
 			]);
 
-			const usersByRole = await roleService.countUsersByRole();
+			const usersByRole = await roleRepository.countUsersByRole();
 
 			expect(usersByRole).toStrictEqual({ admin: 2, member: 3, owner: 1 });
 		});
